[#180] refactor: response DTO에 Schema 어노테이션 적용
